pyproject.toml
[project]
name = "mouldcatalogue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "supabase>=2.15.3",
]
